requirements.txt
Django==5.2.7
psycopg2-binary==2.9.10
gunicorn==23.0.0
whitenoise==6.8.1
python-dotenv==1.0.1
dj-database-url==2.1.0